Add utilities and vue
